consumer/requirements.txt
kafka-python
pymongo
redis
